GH-664 added translation lane
